Function to generate a SummarizedExperiment using a h5ad file for the counts

Function to generate a SummarizedExperiment using a h5ad file for the counts

dataset_h5ad(
  h5ad_file_counts,
  h5ad_file_tpm = NULL,
  cell_id_col = "ID",
  cell_type_col = "cell_type",
  name = "SimBu_dataset",
  spike_in_col = NULL,
  additional_cols = NULL,
  filter_genes = TRUE,
  variance_cutoff = 0,
  type_abundance_cutoff = 0,
  scale_tpm = TRUE
)

Arguments

h5ad_file_counts

(mandatory) h5ad file with raw count data

h5ad_file_tpm

h5ad file with TPM count data

cell_id_col

(mandatory) name of column in Seurat meta.data with unique cell ids

cell_type_col

(mandatory) name of column in Seurat meta.data with cell type name

name

name of the dataset; will be used for new unique IDs of cells#' @param spike_in_col which column in annotation contains information on spike_in counts, which can be used to re-scale counts; mandatory for spike_in scaling factor in simulation

spike_in_col

which column in annotation contains information on spike_in counts, which can be used to re-scale counts; mandatory for spike_in scaling factor in simulation

additional_cols

list of column names in annotation, that should be stored as well in dataset object

filter_genes

boolean, if TRUE, removes all genes with 0 expression over all samples & genes with variance below variance_cutoff

variance_cutoff

numeric, is only applied if filter_genes is TRUE: removes all genes with variance below the chosen cutoff

type_abundance_cutoff

numeric, remove all cells, whose cell-type appears less then the given value. This removes low abundant cell-types

scale_tpm

boolean, if TRUE (default) the cells in tpm_matrix will be scaled to sum up to 1e6

Value

Return a SummarizedExperiment object

Examples

counts <- Matrix::Matrix(matrix(rpois(3e5, 5), ncol=300), sparse = TRUE)
tpm <- Matrix::Matrix(matrix(rpois(3e5, 5), ncol=300), sparse = TRUE)
tpm <- Matrix::t(1e6*Matrix::t(tpm)/Matrix::colSums(tpm))

colnames(counts) <- paste0("cell-",rep(1:300))
colnames(tpm) <- paste0("cell-",rep(1:300))
rownames(counts) <- paste0("gene-",rep(1:1000))
rownames(tpm) <- paste0("gene-",rep(1:1000))

annotation <- data.frame("ID"=paste0("cell-",rep(1:300)), 
                         "cell_type"=c(rep("T cells CD4",50), 
                                       rep("T cells CD8",50),
                                       rep("Macrophages",100),
                                       rep("NK cells",10),
                                       rep("B cells",70),
                                       rep("Monocytes",20)),
                         row.names = paste0("cell-",rep(1:300)))

ad <- anndata::AnnData(X=counts, 
                       obs = data.frame(genes=rownames(counts), row.names=rownames(counts)), 
                       var = annotation)                                                     
counts_file <- tempfile(pattern = 'test',fileext = '.h5ad')
anndata::write_h5ad(anndata = ad, filename = counts_file)

ds_h5ad <- SimBu::dataset_h5ad(h5ad_file_counts = counts_file,
                               name = "h5ad_dataset",
                               cell_id_col = 'ID', 
                               cell_type_col = 'cell_type')
#> Error in generate_summarized_experiment(annotation = annotation, count_matrix = count_matrix,     tpm_matrix = tpm_matrix, name = name, spike_in_col = spike_in_col,     additional_cols = additional_cols, filter_genes = filter_genes,     variance_cutoff = variance_cutoff, type_abundance_cutoff = type_abundance_cutoff,     scale_tpm = scale_tpm): object 'annotation' not found