Build a dataset using a single sfaira entry ID

Build a dataset using a single sfaira entry ID

dataset_sfaira(
  sfaira_id,
  sfaira_setup,
  name,
  spike_in_col = NULL,
  additional_cols = NULL,
  force = F,
  filter_genes = T,
  variance_cutoff = 0,
  type_abundance_cutoff = 0,
  scale_tpm = T
)

Arguments

sfaira_id

(mandatory) ID of a sfaira dataset

sfaira_setup

(mandatory) the sfaira setup; given by setup_sfaira

name

name of the dataset; will be used for new unique IDs of cells

spike_in_col

which column in annotation contains information on spike_in counts, which can be used to re-scale counts

additional_cols

list of column names in annotation, that should be stored as well in dataset object

force

boolean, if TRUE, datasets without annotation will be downloaded, FALSE otherwise (default)

filter_genes

boolean, if TRUE, removes all genes with 0 expression over all samples & genes with variance below variance_cutoff

variance_cutoff

numeric, is only applied if filter_genes is TRUE: removes all genes with variance below the chosen cutoff

type_abundance_cutoff

numeric, remove all cells, whose cell-type appears less then the given value. This removes low abundant cell-types

scale_tpm

boolean, if TRUE (default) the cells in tpm_matrix will be scaled to sum up to 1e6

Value

dataset object