download a specific dataset from sfaira by an ID

download a specific dataset from sfaira by an ID

download_sfaira(
  setup_list,
  id,
  force = F,
  synapse_user = NULL,
  synapse_pw = NULL
)

Arguments

setup_list

the sfaira setup; given by setup_sfaira

id

the ID of the datasets

force

logical; TRUE if you want to force the download, even though no cell-type annotation exists for this dataset. Default if FALSE

synapse_user

character; username for synapse portal (https://www.synapse.org)

synapse_pw

character; password for synapse portal (https://www.synapse.org)

Value

a anndata object, stores counts, metadata and other information on the dataset